Update BelongsToDependency.php
okay line 52 broke the package with `->toBase()` which adds a `->whereNull($this->meta['dependsOnKey'])` to the query. Then when we add the `dependsOnValue` the query basically becomes 
```
$query->whereNull($this->meta['dependsOnKey'])->where($this->meta['dependsOnKey'], $request->dependsOnValue)
```

It cannot be null AND have value at the same time. Since this PR says this fixes Nova v3.19 issue then we are only left with